Update angucomplete.js
In a large dataset occasionally a parameter could be undefined instead of an empty string.  For example {name: undefined, dob: '02/16/1985'} whereas all the other records contain data.  In this case instead of calling toLowerCase() on the undefined variable and throwing an error - simply check if it is a string first and if not it will move onto the next result.